Address @zie1ony feedback on token ownership.
This change changes the ownership for stored contracts from their
contract hash to their contract package. The concern here is that
upgrading a "holder" contract would essentialy lock the tokens, and
therefore make them unusable. With changing the ownership into the
package hash then regardless of which version is calling the ownership
is maintained.